% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split_names.R
\name{get_split_names}
\alias{get_split_names}
\title{Method for extracting names of splitting features used in a tree.}
\usage{
get_split_names(tree, data)
}
\arguments{
\item{tree}{a tree object.}

\item{data}{train or test set.}
}
\value{
names of splitting features extracted from tree object.
}
\description{
Method for extracting names of splitting features used in a tree.
}
